Microsoft-Speech-SDK/requirements.txt
azure-cognitiveservices-speech==1.6.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Naked==0.1.31
prompt-toolkit==1.0.14
Pygments==2.4.2
PyInquirer==1.0.3
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7